requirements.txt
pandas==2.2.2
beautifulsoup4==4.12.3
requests==2.32.2
dlt==0.4.3
harlequin==1.20.0
dbt-core==1.8.1
dbt-duckdb==1.8.0
shandy-sqlfmt[jinjafmt]==0.21.3
yfinance==0.2.40
langchain #==0.2.4
# torch #==2.3.1
# transformers #==4.41.2
accelerate
python-dotenv
langchain-anthropic
streamlit==1.35.0
duckdb==0.10.3
